% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCCP.R
\name{initialize.svm}
\alias{initialize.svm}
\title{Core algorithm for initializing solution}
\usage{
initialize.svm(X, Y, w_neg = 0.5, C = 1)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix.}

\item{Y}{A \eqn{n \times 1} binary vector with value -1 and +1.}

\item{w_neg}{class weight for negative label with value in (0,1).}

\item{C}{inverse of regularized hyperparameter in SVM;}
}
\value{
list res with res$beta and res$b denoting the slope and intercept seperately
}
\description{
Compute initialized linear classifier for later usage
}
