% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCCP.R
\name{dccp}
\alias{dccp}
\title{Core algorithm for improving a linear classifier in terms of some metrics}
\usage{
dccp(
  X,
  Y,
  beta_init,
  b_init,
  gamma,
  psi_k,
  max_iter_num,
  max_rel_gap,
  metrics = "ROC"
)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix.}

\item{Y}{A \eqn{n \times 1} binary vector.}

\item{beta_init}{Slope coefficient for initialized discriminant function.}

\item{b_init}{Intercept coefficient for initialized discriminant function.}

\item{metrics}{Specify the metric to improve for initialized linear classifier.}
}
\value{
beta: Generated slope coefficient

b: Generated intercept coefficient
}
\description{
Compute improved linear classifier to initialized linear classifier under certain metrics.
}
